changed dt type

+++ include/game_module.h
@@ -13,7 +13,7 @@ typedef int (*FGameInit)(GameState*);
 // todo(brandon): consider allowing for only specific implementations and or use dif names. 
 // typedef int (*FGameUpdateA)(int, ScreenData*);
 // todo(brandon): make controller, a list or something so games can be multi controllered.
-typedef int (*FGameUpdate)(int, ScreenData*, GameState*, GameInputController*);
+typedef int (*FGameUpdate)(float, ScreenData*, GameState*, GameInputController*);
 
 typedef int (*FGameShutdown)();
